requirements.txt
coloredlogs==15.0.1
iMail==1.0.0
matplotlib==3.3.0
numpy==1.24.4
pandas==1.3.5
Pillow==10.4.0
PyYAML==6.0.2
scikit_learn==1.3.2
scipy==1.11.1
# torch==2.4.0+cu118
# torchvision==0.19.0+cu118
tqdm==4.66.4
